poprawa zapisu i odczytu z pliku ocen i poprawne generowanie statystyk

+++ MyMountainApp/IMTrails.cs
@@ -9,7 +9,7 @@ namespace MyMountainApp
         string TrailName { get; }
         string TrailDescription { get; }
         string TrailPlace { get; }
-        float LenghtOfTheTrail { get; }
+        int LenghtOfTheTrail { get; }
 
         event GradeAddedDelegate GradeAdded;
